feat: add ttl setting
